requirements.txt
discord.py
requests
websocket-client
Pillow
